+++ packages/signers/raw-tx-creator/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./from-raw-signer"
+export type { TxCreator, SignerTxCreator } from "@polkadot-api/tx-creator"