+++ packages/signers/ledger-signer/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Removed
+
+- Replaced `LedgerTxCreator` for `CommonSignerTxCreator`
+
 ## 0.4.0-rc.2 - 2026-07-14
 
 ### Added