+++ packages/signers/pjs-signer/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Removed
+
+- Replaced `PjsTxCreator` for `CommonSignerTxCreator`
+
 ## 0.8.0-rc.2 - 2026-07-14
 
 ### Added