+++ packages/signers/signers-common/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- Export type `CommonSignerTxCreator` for signer TxCreators with the common enhancers
+
 ## 0.3.0-rc.2 - 2026-07-14
 
 ### Fixed